requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Cors
paho-mqtt